Use image from repository for action (require release branches!)
The action will pull the docker image from dockerhub now, instead of
building it everytime. This hopefully should decrease build times.
The action uses the latest tag in the main branch, so to specify the
right Capella version, it is necessary to create a release branch and
change the tag for the action there.

Testing has also been changed:

On any branch that is not using a slashed prefix (such as release/...):
- First build the image (and tag: latest)
- Test the action (it's using the latest tag by default)